requirements.txt
streamlit==1.22.0
pandas==2.0.3
numpy==1.23.5
transformers==4.31.0
Pillow==9.6.0
